logica y factorizacion de codigo
